got the save feature to somehow work 2
