+++ app/src/main/java/com/example/mspaint/mainui/mainScreen.kt
@@ -1,18 +1,8 @@
 package com.example.mspaint.mainui
 
-import android.content.ContentResolver
-import android.content.ContentValues
-import android.content.Context
-import android.content.Intent
-import androidx.compose.ui.platform.LocalDensity
 import android.content.res.Resources
 import android.graphics.Bitmap
-import android.net.Uri
-import android.os.Build
-import android.os.Environment
-import android.provider.MediaStore
 import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.gestures.detectDragGestures
@@ -24,11 +14,11 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Slider
 import androidx.compose.material3.SliderDefaults
 import androidx.compose.material3.Text
-import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
@@ -39,7 +29,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
@@ -49,20 +38,17 @@ import androidx.compose.ui.graphics.asAndroidBitmap
 import androidx.compose.ui.graphics.drawscope.CanvasDrawScope
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
-import com.example.mspaint.R
+import com.example.mspaint.MainActivity
 import com.example.mspaint.canvasObjectData.PathProperties
 import com.example.mspaint.canvasObjectData.hue
 import com.example.mspaint.canvasObjectData.pencilWidth
+import com.example.mspaint.tools.saveToDisk
 import com.example.mspaint.ui.theme.PureBlack
-import androidx.compose.ui.unit.LayoutDirection
-import com.example.mspaint.MainActivity
-import java.io.File
-import java.io.FileOutputStream
 import kotlin.math.cos
 import kotlin.math.sin
 
@@ -486,40 +472,7 @@ fun MainScreen() {
 
 
 
-private fun Bitmap.saveToDisk(context: Context) {
-    val fileName = "test2.png"
-    val filePath = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES)
-    val directory = File(filePath, "DoddleDoodle")
-    val values = ContentValues()
 
-    if (!directory.exists()) {
-        directory.mkdirs()
-    }
-    val file = File(directory,fileName)
-
-    file.writeBitmap(this, Bitmap.CompressFormat.PNG,100)
-
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-        val contentValues = ContentValues().apply {
-            put(MediaStore.Images.Media.DISPLAY_NAME,file.name)
-            put(MediaStore.Images.Media.MIME_TYPE, "image/png")
-            put(MediaStore.Images.Media.RELATIVE_PATH, "${Environment.DIRECTORY_PICTURES}/${directory.name}")
-        }
-        context.contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI,contentValues)
-    }
-    else {
-        context.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(file)))
-    }
-
-}
-
-
-private fun File.writeBitmap(bitmap: Bitmap, format: Bitmap.CompressFormat, quality: Int) {
-    outputStream().use {out->
-        bitmap.compress(format,quality,out)
-        out.flush()
-    }
-}
 
 private fun Int.dpToFloat(): Float {
    return (this * Resources.getSystem().displayMetrics.density)
